Fix pr notes
